conditional tests fixes
